backend/requirements.txt

Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
python-dotenv==1.0.1
dj-database-url==2.1.0
